+++ packages/band/src/bars/Bar.tsx
@@ -162,7 +162,7 @@ export const Bar = ({
     const valueScale = horizontal ? (scales.x as LinearAxisScale) : (scales.y as LinearAxisScale)
 
     // compute spacings between (possibly grouped) bars
-    const [barWidth, barGap] = getInternalWidthAndGap(indexScale, keys, paddingInternal, stacked)
+    const [barWidth, barGap] = getInternalWidthAndGap(indexScale, keys, paddingInternal, variant)
     const preparedData = useMemo(
         () =>
             prepareData(
@@ -170,7 +170,7 @@ export const Bar = ({
                 indexScale,
                 valueScale,
                 horizontal,
-                stacked || paddingInternal === null,
+                stacked,
                 barWidth,
                 barGap,
                 disabled